+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- 105 Job applications
+- 109 Job alert subscription
+- 110 Job search alert notifications
+- 111 Job alert unsubscribe
+- 133 Search by job number
+
 ## 4.0 (2020-02-24)
 
 Compare: https://git.torchbox.com/buckinghamshire-council/bc/compare/3.2...4.0